csi_automated/requirements.txt
hampel==1.0.2
matplotlib==3.10.1
numpy==2.2.3
pandas==2.2.3
Pillow==11.1.0
pytz==2024.2
scikit_learn==1.6.1
scipy==1.15.2